requirements.txt
Django==1.8.3
PAM==0.4.2
Pillow==2.3.0
PyYAML==3.10
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
amqp==1.3.3
anyjson==0.3.3
apt-xapian-index==0.45
argparse==1.2.1
billiard==3.3.0.15
celery==3.1.6
chardet==2.0.1
cl==0.0.3
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
django-tastypie==0.12.2
djangorestframework==3.2.0
duplicity==0.6.23
html5lib==0.999
httplib2==0.8
ipython==1.2.1
kombu==3.0.7
librabbitmq==1.0.3
lockfile==0.8
lxml==3.3.3
mailer==0.7
numpy==1.8.2
oauthlib==0.6.1
oneconf==0.3.7
openshot==1.4.3
pexpect==3.1
piston-mini-client==0.7.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyinotify==0.9.4
pyparsing==2.0.1
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5ubuntu1
python-dateutil==1.5
python-debian==0.1.21-nmu2ubuntu2
python-memcached==1.53
python-mimeparse==0.1.4
pytz==2012c
pyxdg==0.25
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.5.2
software-center-aptd-plugins==0.0.0
ssh-import-id==3.21
system-service==0.1.6
urllib3==1.7.1
virtualenv==1.11.4
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
